Implementação Parcial da Aula 4 - implementado div e mult

+++ src/aula4/padrao_interpreter/calculadora/Expressao.java
@@ -2,5 +2,5 @@ package aula4.padrao_interpreter.calculadora;
 
 public interface Expressao {
 	
-	int avalia();
+	double avalia();
 }
